requirements.txt
# Florida School Surtax Oversight Dashboard
# Core dependencies

Flask>=2.3.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Database
# SQLite is built into Python, no extra package needed

# Optional: For production deployment
gunicorn>=21.0.0

# Optional: For advanced analytics (uncomment as needed)
# pandas>=2.0.0
# numpy>=1.24.0
# scikit-learn>=1.3.0

# Development dependencies
# pytest>=7.0.0
# pytest-flask>=1.2.0
